fixed habitEvent bugs where adds, edits, deletes had issues
